Fix ordering and placeholder bugs in SettingsViewController.
Ordering:
The order of elements in @IBOutlet collection arrays is not deterministic. However, logic for swapping the buttons on the toolbar depends on consistent ordering among the add buttons, the remove buttons, and the image views. Make the order explicit by assigning index tags to the objects in Interface Builder, then sorting the arrays in the view controller at runtime.

Placeholders:
The add-image completion block was comparing the image view's current image to the placeholder image in order to decide whether to replace the add button with the remove button on the toolbar. However, if the placeholder image was assigned by the Storyboard, not the view controller, then the comparison would fail because they were two different instances of the placeholder image. The toolbar would then incorrectly show the remove button. (Repro: on a clean install, start to add a photo to the last image view, then cancel.)

Instead, encapsulate the placeholder image logic into the view itself, so the Storyboard and the view controller can both use nil to represent an empty image.